[FIRRTL] Add layer verifier, no bind-under-inline
Add a verifier to the `firrtl.layer` op to disallow bind layers under
inline layers.  We can potentially support this in the future, however, it
does not easily work now.  The plan here is to add basic support for
inline layers and then turn on more support as is needed.

To support this in the future would require making the `` `ifdef `` logic
be placed inside the bind file.  To support arbitrary
bind-under-inline-under-bind requires additional thought.  (I do not know
how to handle this latter case today.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>