[FIRRTL] Change Inline Layers to Candidate ABI
Change the emission for Verilog text macros used to enable inline layers
from using a `_` as a delimiter to a `$` as a delimiter.  This is the
current ABI proposal.

This ABI is chosen because it avoids ambiguity with a delimiter that can
occur in FIRRTL modules (for the most part).  FIRRTL does have escaped
identifiers which could use a `$`.  However, this is unlikely.  An
alternative that was rejected would be to use `-` like the bind layer
files do.  However, switching to `-` would require using Verilog escaped
identifiers which seems like both a weird ABI and a pretty annoying one.
This also will run into problems with the existing Verilog emitter which
doesn't properly emit escaped identifiers as they require trailing
whitespace to terminate them.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>