[FIRRTL] Change firrtl-lower-xmr to a walk
Change the way that the `LowerXMR` pass visits operations in modules from
just visiting operations in the body to a walk.  This is done to fix a bug
exposed when compiling inline layers that define probes.  Previously,
operations inside the `sv.ifdef` that this lowering creates are not
visited resulting in a `LowerXMR` pass failure.  This change future-proofs
this pass so that any new operations added to FIRRTL which may include
regions and blocks will do something sane.  This also makes the pass more
relocatable within the FIRRTL pass pipeline.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>