[Python] Remove multiple inheritance from HW helper class.
As we move towards nanobind, the main issue is that nanobind
explicitly does not want to participate in multiple inheritance, which
means we can't use multiple inheritance in OpViews. In the HW
bindings, we mix in the ModuleLike class with multiple inheritance to
share some code between HWModuleOp and HWModuleExternOp.

To avoid multiple inheritance with a little duplication, this changes
the helper methods on ModuleLike to staticmethods, and adds
corresponding instance methods to HWModuleOp and HWModuleExternOp that
delegate to the implementation defined on ModuleLike.