[PyCDE] Add fork, join, and merge channel functions
- <ChannelSignal>.fork creates two new channels, waits until they are
both available, then accepts an input. Also buffer the output channels
to avoid combinational loops.
- Channel.join waits on two channels then creates a message on the one
output channel containing a struct with field 'a' equal to input channel
A's content and likewise for channel B.
- Channel.merge funnels two channels together into a single output
stream.

This is functionality which really should be handled by the DC dialect
but it's not ready for primetime.