[Python] Manually cast enums while we transition to nanobind.
As we transition from pybind11 to nanobind, the enums defined upstream
will be defined by nanobind. This no longer works with automatic
casting in the pybind11 APIs we define ourselves. To work around this
while we smooth the transition, this updates the APIs to work with
plain py::objects and manually cast to the appropriate C API enums in
a way that's compatible with both the pybind11 and nanobind versions
of enums.