[Arc][ArcToLLVM] Fix seq.clock_gate lowering
This op did not properly lower because it created comb ops for which no lowering pattern was registered. Instead, lower to LLVM directly, but also register the CombToArith patterns to support comb as intermediate operations. This is also consistent with registering HWToLLVM and CombToLLVM in this pass directly. Also, add a missing test for the clock gate lowering.