[FIRRTL] Reuse XMRRefOps in LowerXMR
Change the `LowerXMR` pass to reuse previously created `XMRRefOp`s within
a module.  This is done for later, easier processing of these operations
in `LowerToHW`.  While these operations could be made to be CSE'd (by
giving them the `Pure` trait), it is advantageous to instead not create
duplicate operations as this avoids needing to run schedule a mostly
useless CSE.

This is primarily done to allow for Verilator-compatible lowering of
`force_initial` and `release_initial` in `LowerToHW` where it would be
useful to know when a force and release are referring to the same
reference.  (Yes, full alias analysis is necessary to do this correctly
and relying on the CSE of `LowerXMR` is going to be buggy.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>