[circt-test] Allow tests to filter according to test runners
Interpret the `require_runners` and `exclude_runners` attributes on
formal tests as a filter for the potential runners that can execute a
test. If the set of required runners is not empty, a runner from this
set is picked. Tests for which no runner is available report as
"unsupported", for example if none of the available runners match the
filter.