[Python] Switch CIRCT Python extension and dialects to nanobind. (#8110)
This follows the general nanobind porting guide and the upstream MLIR
changes in the same vein. For the most part, this is a very direct
migration accomplished with sed.

There were a couple minor differences.

Nanobind is more strict about implicit casting, which popped up in a
couple places:

* In MSFTModule, getNearestFreeInColumn was expecting a
CIRCTMSFTPrimitiveType (int), but was actually being passed
PrimitiveType enums at the call sites. Nanobind refused to
automatically cast enum to int, so it was done manually.
* In SVModule, SVAttributeAttr.get was expecting a bool for
emitAsComment, even though it was declared as having a default of
None. Nanobind refused to automatically cast None to bool, so it was
done manually, with a default of false.

Nanobind also prints enums slightly differently, so one FileCheck test
had to be updated for ESI.

Otherwise, the changes were purely mechanical.