[FIRRTL][IMCP] Overdefine ports of modules with unknown symbol uses (#8115)
If a module is referenced from an unknown top-level operation, i.e. an
operation that is not an `hw.hierpath`, mark the module's inputs as
overdefined. IMCP cannot reason about how the module is used by such an
unknown operation, and therefore should assume that the operation might
instantiate the module and apply arbitrary values to its input.

As an example, the `firrtl.formal` operation may refer to a private
module as to be executed as a formal test, applying symbolic values to
the module's inputs. While IMCP could simply special-case the
`firrtl.formal` operation, it feels cleaner to make the pass defensive
in the presence of _any_ operation which it does not explicitly know how
to deal with.