[FIRRTL] Add require and ensure intrinsics
Add the `circt.verif.require` and `circt.verif.ensure` intrinsics as a
way to produce the `verif.require` and `verif.ensure` operations from
FIRRTL. At a later point we'll probably want to promote the five verif
ops `assert`, `assume`, `cover`, `require`, and `ensure` to the FIRRTL
spec and merge the functionality of the `circt.verif.assert` intrinsic
with the `firrtl.assert` operation.

Note that this requires dropping of the clock operand from
`verif.require` and `verif.ensure`. This operand was never used, so it
feels like a reasonable change. At a later point we might want to merge
`verif.clocked_assert` into `verif.assert` by adding clock operands to
the verif ops in general.