[FIRRTL] Add contract declaration
Add `contract` declaration support to the FIRRTL parser, and a
corresponding operation to the FIRRTL dialect. A contract acts like a
`node`, but allows for additional formal properties to be expressed in
the contract body. The contract can then either be treated as a node for
synthesis, where the formal properties are just discarded, or it can be
used to subdivide a larger formal problem using a combination of the
assume-guarantee and cutpointing techniques commonly used in formal
verification tools.

This has not landed in the FIRRTL spec yet. As soon as that happens, I
will replace the `missingSpecFIRVersion` use with the actual spec
version.

Actual lowering follows in a separate commit.