[ESI Runtime] Pluggable channel engines
Introduces the concept of 'engines' -- things which are responsible for
transmitting/recieving messages over a channel. Since this is a new
concept (and was previously done only in cosim), add a new section to
the manifest.

Re-use a bunch of stuff which supports services. Hacky, but it works.
The entire manifest thing could use a re-think and second iteration. It
is, however, make-it-work time.