[Arcilator] Allow running verif.simulation ops
Add the `LowerVerifSimulations` pass to the Arc dialect. This pass
converts any `verif.simulation` ops in the input into a test harness
module and a `func.func` that instantiates the module and applies the
necessary stimulus to it using `arc.sim.*` ops.

Also add the pass to the arcilator pipeline. This allows the user to
run any `verif.simulation` ops in the input design by specifying the
`--run --jit-entry=<simulation-name>` command line options. We may want
to streamline this in the future.

Also add an integration test which verifies an adder circuit and prints
the simulation results.