[ImportVerilog] Add option to lower always @* as always_comb (#8271)
Add the `--always-at-star-as-comb` option to circt-verilog and add a
corresponding option to ImportVerilog. When this option is set, detect
`always @*` in the Verilog AST and lower it as a `always_comb`. This is
a common pattern in synthesizers and simulators, since the traiditonal
`always @*` in Verilog does not accurately describe the behaviour of
combinational logic.

Enable this option by default, which seems to be a common behaviour
among existing tools.