[LLHD] Allow processes to yield results
Allow `llhd.process` ops to define result values. Extend the `llhd.wait`
and `llhd.halt` op with an additional list of yield operands which will
be returned as the results of the process when execution suspends. This
will allow us to later factor `llhd.drv` ops out of the process and
instead transport the driven value as a yield operand.