Add mux-of-const canonicalizations to FIRRTL
These canonicalizations are already present in comb, but I am adding
them to the firrtl dialect so that they can run before lower layers.

We generally pool constants at the top of a module. If constants are
used within a layerblock, they can be captured as ports during lower
layers, which in turn prevents comb canonicalizers from running.

Ideally, layer-sink would aggressively clone or sink constants into
layerblocks, but we're not there yet.

These canonicalizations are important because they can drastically
simplify the if/else branches which drive registers.  When lowering a
seq.firreg to sv, we incorporate any muxes feeding into the register as
if-else branches, which conditionally drive the register. These
canonicalizations can transform a deeply nested mux chain into a tree of
and/or expressions, which reduces the depth of if/else branches