[circt-test] fix SymbiYosys integration test (#7886)
#7756 adds the SymbiYosys test runner and requires the `sby` feature, 
but did not add the corresponding available feature in `lit.cfg.py`, so 
this test never seems to run, and #7763 introduces a syntax error, a 
backslash in the f-string expression part, in the Python script.

Unlike #7884, there is no need to modify PATH variable for the script 
since `sby` should be in the PATH if we can check it in CMake.

Signed-off-by: unlsycn <unlsycn@unlsycn.com>