[LLHD] Hoist drives out of processes
Extend the `HoistSignals` pass and hoist drives before halt and wait
terminators out of the parent process. All drives that have no
side-effecting operations in between them and the block's suspending
terminator can be hoisted. If the same signal is driven multiple times
in the same block, only the last drive is hoisted.

To hoist drives, the pass collects all drives for each driven signal
slot and creates new yield operands and process results to carry the
drive operands out of the process. Special care is taken to not create
unnecessary results for values that are uniform across all drives and
available outside the processes.