[FIRRTL] Add simulation op and parser support
Add the `firrtl.simulation` op, which declares a simulation unit test
and references a module to act as the top-level for the simulation. The
target module must have exactly four ports, clock, init, done, and
success, with corresponding directions and types. This new op is
analogous to the `firrtl.formal` op.

Also add support for a `simulation` construct in the parser, also
analogous to the `formal` construct.

A later commit will add a lowering to `verif.formal`.