[FIRRTL][LowerLayers] Emit bind files for each module
For each public module, emit a bind file for every layer we know about,
regardless of whether the layer has actually been used under the module.

For each private module, emit a bind file for every layer used under that
module, whether directly referenced via a layerblock op, or indirectly through
an instance op. Do not emit a bind file for a layer, if the layer is not used.

For a bind file emitted for a module M and layer L, the bind file should
automatically enable any parent layers of L. If L has a parent layer L', and L'
is a bind layer, include the bind file for M/L'. M is guaranteed to have a bind
file for L'.  If L' is a bind layer, define the macro for L' and enable any
parents of L'.

The bind file should also enable the layer globally. If M instantiates a module
M', include the bind file for M'/L, if it exists.

Use structured sv ops to build the bind file contents.