[FIRRTL] Add special printf substitutions
Add ODS descriptions of "special substitutions" for printfs.  This is
intended to represent a new substitution style for FIRRTL printfs that is
double curly brace-delimited.  E.g.:

    printf(clock, en, "[{{SimulationTime}}] hello world")

The plan for this is to have an enum representation for legal
substitutions and to parse out the special substitutions to avoid both
complicated verification of the legality of a format string and
difficulty for passes to understand special substitutions.  (We may want
to transform special substitutions, like simulation time, and it is
tedious if the IR storage for this constantly has to be parsed.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>