[FIRRTL] Parse/emit special printf substitutions
Add parser and emitter support for new "special substitutions".  These are
substitutions that take the form of `{{Name}}` where "Name" is a list of
keywords defined by the FIRRTL spec.  This is intended to be used to
represent "simulation time" and "hierarchical path" and will be compiled
to `$time` and `%m` in Verilog.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>