[FIRRTL] Add FormatStringType, TimeOp Expression
Add a new type, `!firrtl.fstring.time`, and a single expression that can
create it: TimeOp / `firrtl.fstring.time`.  These are used to model the
notion of simulation time that should be emitted into a format string.

Allow for this to be a legal substitution operand to the existing
PrintfOp.  Add printing/parsing support for this which will convert the
special substitution `{{SimulationTime}}` into a `TimeOp` and an added
substitution at the correct location.  Add emission support which inverts
this representational conversion.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>

[FIRRTL] Parse/emit special printf substitutions

Add parser and emitter support for new "special substitutions".  These are
substitutions that take the form of `{{Name}}` where "Name" is a list of
keywords defined by the FIRRTL spec.  This is intended to be used to
represent "simulation time" and "hierarchical path" and will be compiled
to `$time` and `%m` in Verilog.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>