[FIRRTL] Add fprintf operation
Add support for fprintf operation to FIRRTL dialect. This operation is similar
to printf but writes to a file instead of stdout. The output file name is
interpreted as a format string in SystemVerilog, allowing users to use %m to
create unique file names per instance.

The operation has the following syntax:
fprintf(clock, cond, "output_file", "format_string", substitutions...)