[FIRRTL] Fold regreset to its reset value iff the type matches
If the regreset has a constantly-high reset, then we can replace the
register with its reset value. However, if the reset value's type does
not match the type of the register, we cannot do the replacement.

This commit adds a check to an old canonicalizer, to ensure the reset
value's type matches, before doing the replacement.