[Seq] Fix width errors in HWMemSimImpl
Fix some width errors which could occur in the HWMemSimImpl pass.  The
problem is that a loop variable may need to be sized one bit larger than
what is used for its address in order for it to hit its loop bound.
However, if this is sized large enough to ever reach the loop bound, it is
too large for the index.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>