[LLHD] Make probes side-effect-free in graph regions
Make `llhd.prb` ops only have read side effects if used in an SSACFG
region. This works around an issue where CSE would not move or eliminate
probes across drives and other processes, even though that would be
legal in a graph region. In graph regions, a probe acts like a pure op
dereferencing a signal and the exact order of operations, especially the
order with respect to drives and other ops interacting with the signal
do not matter.