[FIRRTL] When replacing register with reset value, attemp coercion
We have a canonicalization which can replace a register with its reset
value, if the reset signal is constantly high. However, the reset-value
of a register does not have to exactly match the type of the register,
it only has to be equivalent. Previously, this canonicalization would
fail if the reset value's type doesn't exactly match the register's
type.  This commit makes the canonicalization attempt to coerce the
reset value to the type of the register.

The coercion code will perform sign extension or truncation for
integers, and will recursively attempt to coerce the elements of a
bundle or vector.